# feat: use foreach loop in `GetPrintedWidth` to prevent `CharEnumerator` allocations