@TimothyMakkison 2025-02-16
Don't know if this is a rider thing, but 0cda407 won't let me build due to hundreds of errors.