@belav 2025-02-17
Which SDKs do you have installed locally? At work we ran into the problem that the dotnet analyzers can change between feature releases. Which led to inconsistent builds between devs and build servers. I'm running 9.0.200, the GH action also used that version.

I'm also using Rider and don't have the problem. You could set those two analyzers to `none` in the editorconfig.