@TimothyMakkison 2025-02-18
Thanks for the suggestion, I'm running 9.0.100 and I probably should have tried restarting rider :sweat_smile:. Until then I've disabled the errors via `build.props`